jabber unsigned long -> uint32 patch by Konrad Karczewski [amd64 compatibility]

+++ plugins/jabber/digest.c
@@ -30,12 +30,12 @@ A million repetitions of "a"
 #include <string.h>
 
 typedef struct {
-    unsigned long state[5];
-    unsigned long count[2];
+    uint32_t state[5];
+    uint32_t count[2];
     unsigned char buffer[64];
 } SHA1_CTX;
 
-static void SHA1Transform(unsigned long state[5], unsigned char buffer[64]);
+static void SHA1Transform(uint32_t state[5], unsigned char buffer[64]);
 static void SHA1Init(SHA1_CTX* context);
 static void SHA1Update(SHA1_CTX* context, unsigned char* data, unsigned int len);
 static void SHA1Final(unsigned char digest[20], SHA1_CTX* context);
@@ -63,12 +63,12 @@ static void SHA1Final(unsigned char digest[20], SHA1_CTX* context);
 
 /* Hash a single 512-bit block. This is the core of the algorithm. */
 
-static void SHA1Transform(unsigned long state[5], unsigned char buffer[64])
+static void SHA1Transform(uint32_t state[5], unsigned char buffer[64])
 {
-unsigned long a, b, c, d, e;
+uint32_t a, b, c, d, e;
 typedef union {
     unsigned char c[64];
-    unsigned long l[16];
+    uint32_t l[16];
 } CHAR64LONG16;
 CHAR64LONG16* block;
 #ifdef SHA1HANDSOFF
@@ -156,7 +156,7 @@ unsigned int i, j;
 
 static void SHA1Final(unsigned char digest[20], SHA1_CTX* context)
 {
-unsigned long i, j;
+uint32_t i, j;
 unsigned char finalcount[8];
 
     for (i = 0; i < 8; i++) {
